Refactore Entidad ADM_IMPUESTOS_DOCUMENTOS_OP

+++ Convertidor/Data/EntitiesDestino/ADM/ADM_IMPUESTOS_DOCUMENTOS_OP.cs
@@ -8,10 +8,10 @@ public class ADM_IMPUESTOS_DOCUMENTOS_OP
     public int TIPO_RETENCION_ID { get; set; }
     public int TIPO_IMPUESTO_ID { get; set; }
     public string PERIODO_IMPOSITIVO { get; set; } = string.Empty;
-    public int BASE_IMPONIBLE { get; set; }
-    public int MONTO_IMPUESTO { get; set; }
-    public int MONTO_IMPUESTO_EXENTO { get; set; }
-    public int MONTO_RETENIDO { get; set; }
+    public decimal BASE_IMPONIBLE { get; set; }
+    public decimal MONTO_IMPUESTO { get; set; }
+    public decimal MONTO_IMPUESTO_EXENTO { get; set; }
+    public decimal MONTO_RETENIDO { get; set; }
     public string EXTRA1 { get; set; } = string.Empty;
     public string EXTRA2 { get; set; } = string.Empty;
     public string EXTRA3 { get; set; } = string.Empty;
